Doing an overhaul of the data structuring
There are way too many separate arrays. I am moving them all into a
dictionary for now. It might be better to restructure this as a class
eventually, but it might be too big a step for the group right now.
