+++ IS2_velocity/__init__.py
@@ -1,4 +1,5 @@
-from . import readers
+from . import readers_atl06
+from . import readers_atl03
 from . import correlation_processing
 from . import extract_alongtrack
 from . import atl03_reprocessing
